[MIGR] mrp: new API except mrp_production
MRP module is migrated to the new API. It excludes mrp models that will change
quit a lot during the new MRP conversion. Actually most of the code is not
migrated, but at least small models and modifications are already migrated.
This will lead to cleaner commits for the mrp-only changes.