[FIX] account: avoid quadratic complexity when adding many move lines
The fields journal_id and date are related, stored and required (which should be
changed in master). The value must be set in create ('not null' constrain) even
if it is supposed to be computed after the INSERT.
However, setting a value for a related field will force the update on the linked
record. This makes chained recomputation and reduces the performances of create.

In the case of creating a account.move.line (specifically, creating all the move
lines at the end of the point of sale session, linked to the same account move),
a value for journal_id and date is forced in the override of the `create` method.
Setting a value for it will force the update of the field date and journal of
the move_id which will force the recompute of the journal_id and date of all the
lines for the move_id.
Adding many lines to an account.move will then take more and more time as the
number of lines to recompute increase.

Set a custom attribute on the field to avoid the recomputation and makes the
creation (almost) linar.
Use a custom attribute in stable version to avoid introducing regression in the
orm for a scenario that should not have happen.
The required attribute will be removed and replaced by a python constrain.