[FIX] web_tour: positioning tip & overlay
The jquery-ui position(...) implementation is wrong when the element
to position is animated. So, for the tip to be positioned at the exact
correct position, the animation has to be stopped while the _reposition
is running.

Also, when the tip is opened simply pausing the animation is not right
because the tip could be at a position where there is a space between
the tip and its anchor. So instead of pausing the animation, it has to
be stopped.

The overlay was also wrongly positioned by the size of the tip border
(as the border is 2px wide, placing the overlay at -10px places it in
fact at -8px outside the border). To counter this effect, while being
able to modify the overlay position in JS without caring about that,
the solution was to add a negative margin around the overlay of the
size of the tip border.