[IMP] payment_stripe: improve content on payment page
Show the logo, merchant name, description, currency and amount on the stripe
checkout page.
The logo is a square image that can be provided on Stripe configuration (can not
use the company logo as there is no guarantee of ratio).
More info on the image at https://stripe.com/docs/checkout

Closes 12378